+++ src/Clusterers/Seeders/PlusPlus.php
@@ -6,6 +6,7 @@ use Rubix\ML\Datasets\Dataset;
 use Rubix\ML\Kernels\Distance\Distance;
 use Rubix\ML\Kernels\Distance\Euclidean;
 use Rubix\ML\Specifications\DatasetIsNotEmpty;
+use Rubix\ML\Exceptions\RuntimeException;
 
 use function count;
 
@@ -49,11 +50,17 @@ class PlusPlus implements Seeder
      *
      * @param Dataset $dataset
      * @param int $k
+     * @throws RuntimeException
      * @return list<list<string|int|float>>
      */
     public function seed(Dataset $dataset, int $k) : array
     {
         DatasetIsNotEmpty::with($dataset)->check();
+
+        if ($k > $dataset->numSamples()) {
+            throw new RuntimeException("Cannot seed $k clusters with only "
+                . $dataset->numSamples() . ' samples.');
+        }
 
         $centroids = $dataset->randomSubset(1)->samples();