+++ tests/BootstrapAggregatorTest.php
@@ -7,7 +7,6 @@ use Rubix\ML\DataType;
 use Rubix\ML\Estimator;
 use Rubix\ML\Persistable;
 use Rubix\ML\EstimatorType;
-use Rubix\ML\Backends\Serial;
 use Rubix\ML\Datasets\Unlabeled;
 use Rubix\ML\BootstrapAggregator;
 use Rubix\ML\Regressors\RegressionTree;
@@ -115,8 +114,6 @@ class BootstrapAggregatorTest extends TestCase
      */
     public function trainPredict() : void
     {
-        $this->estimator->setBackend(new Serial());
-
         $training = $this->generator->generate(self::TRAIN_SIZE);
         $testing = $this->generator->generate(self::TEST_SIZE);