+++ tests/CrossValidation/KFoldTest.php
@@ -3,7 +3,6 @@
 namespace Rubix\ML\Tests\CrossValidation;
 
 use Rubix\ML\Parallel;
-use Rubix\ML\Backends\Serial;
 use Rubix\ML\CrossValidation\KFold;
 use Rubix\ML\Datasets\Generators\Blob;
 use Rubix\ML\CrossValidation\Validator;
@@ -53,8 +52,6 @@ class KFoldTest extends TestCase
         $this->estimator = new GaussianNB();
 
         $this->validator = new KFold(10);
-
-        $this->validator->setBackend(new Serial());
 
         $this->metric = new Accuracy();
     }