+++ tests/CrossValidation/MonteCarloTest.php
@@ -3,7 +3,6 @@
 namespace Rubix\ML\Tests\CrossValidation;
 
 use Rubix\ML\Parallel;
-use Rubix\ML\Backends\Serial;
 use Rubix\ML\Datasets\Generators\Blob;
 use Rubix\ML\CrossValidation\Validator;
 use Rubix\ML\CrossValidation\MonteCarlo;
@@ -53,8 +52,6 @@ class MonteCarloTest extends TestCase
         $this->estimator = new GaussianNB();
 
         $this->validator = new MonteCarlo(3, 0.2);
-
-        $this->validator->setBackend(new Serial());
 
         $this->metric = new Accuracy();
     }