+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+- 2.5.0
+    - Added Vantage Point Spatial tree
+    - Blob Generator can now `simulate()` a Dataset object
+    - Added Wrapper interface
+    - Plus Plus added check for min number of sample seeds
+    - LOF prevent div by 0 local reachability density
+
 - 2.4.1
     - Sentence Tokenizer fix Arabic and Farsi language support
     - Optimize online variance updating